Move integration tests to aries-vcx crate
