+++ aries_vcx/tests/utils/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod devsetup_agent;
